Add sub-tasks and system-tasks to Todo list
The commit introduces sub-task feature to the existing tasks and newly added system specific tasks like 'inbox'. Created new SQL migrations to modify the 'Task' table - added new columns such as 'parentId', 'systemId', 'userId', 'completion_date' and 'resolvable'. parentId is used to relate tasks to their sub-tasks. System tasks are identifiable by their systemId. New tasks by default belong to the 'userId' of 'no-user'. Task completion can be scheduled with the 'completion_date'. And, not every task is resolvable, which is determined by the 'resolvable' field. The new structure enhances the application's capability to handle complex task management situations.

In addition to these changes, added typescript configuration for Prisma to interact with the new columns in 'Task'. Adapted the new changes in how tasks are handled in the components. Necessary updates are made in relevant server actions and react-query utility functions to accept and pass required arguments.

A new SQL seed file is introduced to pre populate the system tasks into the database.

Packages are updated and added as required in pnpm-lock.yaml meaning the server was tested and is well functioning with the newly added features. Changes were made with backwards compatibility in consideration.

Tailwind CSS is updated to support better visuals for displaying the new features.

And finally some redundant useEffects are removed making the application more performant.