Add IDEA project configuration files to version control
The commit adds various project configuration files which are specific to the JetBrains IntelliJ IDEA, a popular IDE often used for software development.

These changes help to ensure a unified development environment for all team members, making the project setup easier for new contributors and maintaining consistency in elements such as code styles and formatting across the team.

In addition, specific configurations for tasks such as linting (Eslint, TsLint) and formatting (Prettier) are included, as well as configurations for integration with Git and project-specific settings, such as database connections.

The '.idea' directory was removed from .gitignore to allow these files to be tracked by Git.