Refactor database structure and sign out logic
This commit includes several changes related to the application's database structure and user sign-out process.

The 'init.sql' file was removed as the database schema is are no longer necessary. The 'authCookie' function in 'client.ts' was changed to a 'default export' instead of 'named export'. This offers flexibility during imports.

An extra field named 'root' is added to 'systemIds' object in 'tasks.ts'. In 'actions.ts', modification has been made to the return statement of 'findMany' function for better readability and consistency. Also, 'orderBy' has been added in findSystemTasks function to sort tasks in ascending order by id.

The function 'removeRefreshToken' is introduced which removes the refresh token for a specified user. Also, in prisma/seed.ts file we introduced a new root task entry and assigned this as the parent id of the inbox task.

Changes in '.env' file modifies the schema spelling used in database connection URL.

Created a new hook called 'useUser' which fetch user information and provides functionality for sign out operation. Also, added a new 'SignOutPage' component where the actual sign-out operations are implemented using the 'useUser' hook.