Update dependencies to their latest versions
Several dependencies in the package.json and pnpm-lock.yaml files were updated to get the latest patches, improvements, and bug fixes. These include the "@fortawesome" packages and "@types/node" which are updated to enhance performance and security. The updates ensure that the project stays up-to-date with the latest features and improvements.