Refactor FontAwesome to use CSS classes
Removed the imported FontAwesome icons and switched to using CSS classes directly in line with best practices. This change reduces the size of the bundled JavaScript file, hence enabling faster load times. Additionally, it makes the code neater and more maintainable as it eliminates verbosity. To accomplish this, the FontAwesome CSS file was added to the 'development/css' directory and is linked in 'development/css/index.html'. Lastly, the 'CloseSidebarButton.tsx' file was updated to reflect these changes.